+++ src/AllowedSort.php
@@ -7,7 +7,7 @@ use Spatie\QueryBuilder\Sorts\SortsField;
 use Spatie\QueryBuilder\Enums\SortDirection;
 use Spatie\QueryBuilder\Sorts\Sort as SortClass;
 
-class Sort
+class AllowedSort
 {
     /** @var string */
     protected $sortClass;